% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protOutliersAndMeans.R,
%   R/protOutliersAndMeansOld.R
\name{boxplotMod}
\alias{boxplotMod}
\title{================================================================
boxplotMod to identify outliers in a vector
Outliers that are 3 times the interquartile range from either
quartile; note that the default for boxplots is 1.5 times the
interquartile range}
\usage{
boxplotMod(x, reject.vec = NULL, range = 3)

boxplotMod(x, reject.vec = NULL, range = 3)
}
\arguments{
\item{x:}{a vector of numbers}

\item{reject.vec.i:}{pre-specified rejection locations}

\item{range:}{pre-specified times of IQR range}
}
\description{
================================================================
boxplotMod to identify outliers in a vector
Outliers that are 3 times the interquartile range from either
quartile; note that the default for boxplots is 1.5 times the
interquartile range

================================================================
boxplotMod to identify outliers in a vector
Outliers that are 3 times the interquartile range from either
quartile; note that the default for boxplots is 1.5 times the
interquartile range
}
