% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{cpaSetup}
\alias{cpaSetup}
\title{Set up reference profiles for constrained proportional assignment}
\usage{
cpaSetup(protProfileSummary, markerList = markerList, numDataCols)
}
\arguments{
\item{protProfileSummary}{data frame of protein identifiers (protName) and their relative abundance in centrifugation fractions.}

\item{markerList}{List of reference proteins and their subcellular locations}

\item{numDataCols}{Number of channels of abundance levels}
}
\value{
A matrix markerProfiles giving the abundance level profiles of the subcellular locations
}
\description{
Set up reference profiles for constrained proportional assignment
}
