% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{fitCPA}
\alias{fitCPA}
\title{Carry out constrained proportional assignment on all proteinss}
\usage{
fitCPA(
  protProfileSummary,
  markerProfiles,
  numDataCols = numDataCols,
  log2Transf = F,
  eps = 2^(-5),
  maxit = 10000,
  assignProbsStart = NULL
)
}
\arguments{
\item{protProfileSummary}{data frame of protein names (row names) and relative abundance levels.}

\item{markerProfiles}{A matrix giving the abundance level profiles of the subcellular locations}

\item{numDataCols}{Number of channels of abundance levels}

\item{log2Transf}{Transform data before running CPA? Default is F}

\item{eps}{Small quantity to add to data prior to taking a log transformation; ignored
if log2Transf=F}

\item{maxit}{maximum number of iterations}

\item{assignProbsStart}{A matrix of starting values, one for each protein. The first column must be protName}

\item{maxR}{A value to truncate values of markerProfiles; if NULL (default), ignore}
}
\value{
assignProbsOut  Data frame of proportionate assignments of each protein to compartments
}
\description{
Carry out constrained proportional assignment on all proteinss
}
