% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protOutliersAndMeans.R
\name{geneProfileEstimate}
\alias{geneProfileEstimate}
\title{================================================================
geneProfileEstimate to calculate mean and SE for each channel in
each gene using random effect model or arithmetic mean
random effect model can avoid dominance of a sequence with
too many spectra}
\usage{
geneProfileEstimate(
  protsCombineCnew,
  outlier,
  refCol,
  nCol,
  eps,
  dataUse,
  save.data = FALSE
)
}
\arguments{
\item{protsCombineCnew:}{a matrix of numbers}

\item{outlier:}{outlier-removing method: 1)boxplot,
2)score, 3)none}

\item{refCol:}{number of columns before Mass
Spectrometry data columns}

\item{nCol:}{how many columns in MS data}

\item{eps:}{epsilon to avoid log(0)}

\item{dataUse:}{name of data set}

\item{save.data:}{save output data externally?}
}
\value{
: calculated mean and SE, raw means
                          transformed back (inverse of log2eps)
}
\description{
================================================================
geneProfileEstimate to calculate mean and SE for each channel in
each gene using random effect model or arithmetic mean
random effect model can avoid dominance of a sequence with
too many spectra
}
\examples{
gProfSummaryTMTms2 <- geneProfileEstimate(protsCombineCnew = protsCombineCnew, outlier="boxplot",
refCol=4, nCol=9, eps=0.029885209, dataUse="tmtMS2", save.data=F)
head(gProfSummaryTMTms2)
================================================================
}
