% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protOutliersAndMeans.R,
%   R/protOutliersAndMeansOld.R
\name{geneProfileSummary}
\alias{geneProfileSummary}
\title{================================================================
geneProfileSummary to calculate mean and SE for each channel in
each gene using random effect model or arithmetic mean
random effect model can avoid dominance of a sequence with
too many spectra}
\usage{
geneProfileSummary(
  protsCombineCnew,
  Outlier,
  refCol,
  nCol,
  eps,
  dataUse,
  save.data = TRUE
)

geneProfileSummary(
  protsCombineCnew,
  Outlier,
  refCol,
  nCol,
  eps,
  dataUse,
  save.data = TRUE
)
}
\arguments{
\item{protsCombineCnew:}{a matrix of numbers}

\item{outlier:}{outlier-removing method: 1)boxplot,
2)score, 3)none}

\item{refCol:}{number of columns before Mass
Spectrometry data columns}

\item{nCol:}{how many columns in MS data}

\item{eps:}{epsilon to avoid log(0)}

\item{dataUse:}{name of data set}

\item{save.data:}{save output data externally?}
}
\description{
================================================================
geneProfileSummary to calculate mean and SE for each channel in
each gene using random effect model or arithmetic mean
random effect model can avoid dominance of a sequence with
too many spectra

================================================================
geneProfileSummary to calculate mean and SE for each channel in
each gene using random effect model or arithmetic mean
random effect model can avoid dominance of a sequence with
too many spectra
}
