% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protOutliersAndMeans.R,
%   R/protOutliersAndMeansOld.R
\name{log2eps}
\alias{log2eps}
\title{================================================================
log2eps to log transformation with epsilon adjustment}
\usage{
log2eps(x, .eps = eps)

log2eps(x, .eps = eps)
}
\arguments{
\item{x:}{a vector of numbers}

\item{eps:}{little value to avoid log2(0)}
}
\description{
================================================================
log2eps to log transformation with epsilon adjustment

================================================================
log2eps to log transformation with epsilon adjustment
}
