% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerProfilePlot.r
\name{markerProfilePlot}
\alias{markerProfilePlot}
\title{Plot profiles of reference proteins}
\usage{
markerProfilePlot(
  refLoc,
  markerList = markerListJadot,
  profile,
  refLocationProfiles,
  refProtPlot = NULL
)
}
\arguments{
\item{refLoc}{the name of the reference subcellular compartment to plot}

\item{markerList}{List of reference proteins}

\item{profile}{data frame of protein names (as row names) and relative abundance levels.}

\item{refLocationProfiles}{A matrix refLocationProfiles giving the abundance level profiles of the subcellular locations}
}
\description{
This function plots profiles of reference proteins and also the average profile for each compartment
}
