% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureAreaError.r
\name{mixtureAreaError}
\alias{mixtureAreaError}
\title{Compute area-based error for mixture of two compartment profiles}
\usage{
mixtureAreaError(
  mixProtiProtjProp,
  NstartMaterialFractions = 6,
  Loc1,
  Loc2,
  mix.df
)
}
\arguments{
\item{mixProtiProtjProp}{data frame of CPA estimated proportions for each mixture}

\item{NstartMaterialFractions}{Number of fractions that comprise the starting material}

\item{Loc1}{row number of subcellular location 1 of mixture}

\item{Loc2}{row number of subcellular location 2 of mixture}

\item{mix.df}{true mixture proportions used to make the  mixtures (from proteinMix)}

\item{errorReturn}{Return area of error region if true}
}
\description{
Compute area-based error for mixture of two compartment profiles
}
