% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureHeatMap.r
\name{mixtureHeatMap}
\alias{mixtureHeatMap}
\title{plot mixture of all two compartment profile combinations as panel; assumes eight compartments}
\usage{
mixtureHeatMap(Acup = AcupMarkers, totProt, eps = 0.01)
}
\arguments{
\item{Acup}{relative amount of a given cellular compartment protein that ends up in a given centrifugation fraction}

\item{totProt}{vector of amounts starting material in each fraction}

\item{eps}{small positive constant to add before taking a log transformation}
}
\description{
plot mixture of all two compartment profile combinations as panel; assumes eight compartments
}
