% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixturePlot.r
\name{mixturePlot}
\alias{mixturePlot}
\title{plot mixture of two compartment profiles}
\usage{
mixturePlot(
  mixProtiProtjProp,
  NstartMaterialFractions = 6,
  Loc1,
  Loc2,
  xaxisLab = T,
  yaxisLab = T
)
}
\arguments{
\item{mixProtiProtjProp}{data frame of CPA estimated proportions for each mixture}

\item{NstartMaterialFractions}{Number of fractions that comprise the starting material}

\item{Loc1}{row number of the matrix Acup for compartment 1}

\item{Loc2}{row number of the matrix Acup for compartment 2}

\item{type}{normalized specific amounts or relative specific amounts}
}
\description{
plot mixture of two compartment profiles
}
