% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixturePlotPanel.r
\name{mixturePlotPanel}
\alias{mixturePlotPanel}
\title{plot mixture of all two compartment profile combinations as panel; assumes eight compartments}
\usage{
mixturePlotPanel(
  Acup = AcupMarkers,
  totProt,
  fitType = "relAmt",
  log2Transf = F
)
}
\arguments{
\item{Acup}{relative amount of a given cellular compartment protein that ends up in a given centrifugation fraction}

\item{totProt}{vector of amounts starting material in each fraction}

\item{fitType}{use (normalized) specific amounts or relative specific amounts}

\item{log2Transf}{use log2-transformed values}

\item{type}{normalized specific amounts or relative specific amounts}
}
\description{
plot mixture of all two compartment profile combinations as panel; assumes eight compartments
}
