% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneDistanceCompute.R
\name{nearestGenes}
\alias{nearestGenes}
\title{Compute distances between a particular gene or profile and all other genes, and list the nearest ones}
\usage{
nearestGenes(
  geneName,
  n.nearest = 5,
  distGenes = distUse,
  geneNames = genesUse
)
}
\arguments{
\item{geneName}{Name of gene to which distances are to be computed}

\item{n.nearest}{Number of nearest genes to list}

\item{distGenes}{distance matrix created by, for example, dist(geneProfileSummaryTMTms2[,2:10])}

\item{geneNames}{A list of all genes in a dataset such as geneProfileSummaryTMTms2[,1]}
}
\value{
List of the genes in geneName closest to geneName or to the profile
}
\description{
Compute distances between a particular gene or profile and all other genes, and list the nearest ones
}
\examples{
distUse <- dist(geneProfileSummaryTMTms2[,2:10], method="euclidean")
genesUse <- geneProfileSummaryTMTms2[,1]
nearestGenes("AADAC", n.nearest=10,  distGenes=distUse, geneNames=genesUse)
}
