% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protDistanceCompute.R
\name{nearestProts}
\alias{nearestProts}
\title{Compute distances between a particular gene or profile and all other genes, and list the nearest ones}
\usage{
nearestProts(
  protName,
  n.nearest = 5,
  distProts = distUse,
  protNames = protsUse
)
}
\arguments{
\item{protName}{Name of protein to which distances are to be computed}

\item{n.nearest}{Number of nearest proteins to list}

\item{distProts}{distance matrix created by, for example, dist(protProfileSummaryTMTms2[,2:10])}

\item{protNames}{A list of all proteins in a dataset such as protProfileSummaryTMTms2[,1]}
}
\value{
List of the proteins in protName closest to protName or to the profile
}
\description{
Compute distances between a particular gene or profile and all other genes, and list the nearest ones
}
\examples{
distUse <- dist(protProfileSummaryTMTms2[,2:10], method="euclidean")
protsUse <- protProfileSummaryTMTms2[,1]
nearestProts("AADAC", n.nearest=10,  distProts=distUse, protNames=protsUse)
}
