% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protOutliersAndMeans.R
\name{OutlierIdentify}
\alias{OutlierIdentify}
\title{================================================================
OutlierIdentify to identify outliers in a matrix
using both scoresMod and boxplotMod}
\usage{
OutlierIdentify(protClass, refCol, nCol, save.data = TRUE, byPeptide = FALSE)
}
\arguments{
\item{input.Data:}{a matrix of numbers}

\item{refCol:}{number of columns before Mass Spectrometry
data columns}

\item{nCol:}{how many columns in MS data}

\item{save.data:}{save output data externally?}

\item{byPeptide:}{outliers by peptide(T) or gene(F)?}
}
\description{
================================================================
OutlierIdentify to identify outliers in a matrix
using both scoresMod and boxplotMod
}
