% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamassKmeansRoutines2.r
\name{PrecisionRecallCompute}
\alias{PrecisionRecallCompute}
\title{Compute precision and recall for assignments to subcellular compartments relative to reference}
\usage{
PrecisionRecallCompute(assignVec, compareVec, noneName = "unclassified")
}
\arguments{
\item{assignVec}{list of assignments to subcellular compartments}

\item{compareVec}{list of reference ("true") assignments}

\item{nonName}{name given for genes with no assignment; must come alphabetically last}
}
\value{
cross-classification table of assigned vs comparison locations; then precision and recall
}
\description{
Compute precision and recall for assignments to subcellular compartments relative to reference
}
