% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{proLocAll}
\alias{proLocAll}
\title{Carry out constrained proportional assignment on all genes}
\usage{
proLocAll(
  geneProfileSummary,
  markerLocR,
  n.channels = n.channels,
  log2Transf = F,
  maxit = 10000,
  assignProbsStart = NULL
)
}
\arguments{
\item{geneProfileSummary}{data frame of protein names and relative abundance levels.}

\item{markerLocR}{A matrix giving the abundance level profiles of the subcellular locations}

\item{n.channels}{Number of channels of abundance levels}

\item{maxit}{maximum number of iterations}

\item{assignProbsStart}{A matrix of starting values, one for each gene. The first column must be geneName}
}
\value{
assignProbsOut  Data frame of proportionate assignments of each protein to compartments
}
\description{
Carry out constrained proportional assignment on all genes
}
