% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamassKmeansRoutines2.r
\name{protCluster}
\alias{protCluster}
\title{Assign genes to locations using metamass procedure}
\usage{
protCluster(
  geneProfileSummary,
  refLocProteins,
  refCol = 1,
  nCol = 9,
  clustSize = 5
)
}
\arguments{
\item{geneProfileSummary}{data frame of protein names and relative abundance levels.}

\item{refLocProteins}{List of reference proteins and their subcellular locations}

\item{nCol}{number of columns of abundance levels}

\item{clustSize}{average cluster size}
}
\value{
assignProbsOut  assignments of each protein to compartments
}
\description{
Assign genes to locations using metamass procedure
}
