% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamassKmeansRoutines2.r
\name{protClusterStable}
\alias{protClusterStable}
\title{Assign genes to locations using metamass procedure; Carries out protCluster multiple
  times to assess consistency of assignments to compartments}
\usage{
protClusterStable(
  geneProfileSummary,
  refLocProteins,
  refCol = 1,
  nCol = 9,
  clustSize = 5,
  nReps = 100,
  assignCutoff = 0.5
)
}
\arguments{
\item{geneProfileSummary}{data frame of protein names and relative abundance levels.}

\item{refLocProteins}{List of reference proteins and their subcellular locations}

\item{nCol}{number of columns of abundance levels}

\item{clustSize}{average cluster size}
}
\value{
assignProbsOut  assignments of each protein to compartments, and maximum (most common) assignment
}
\description{
Assign genes to locations using metamass procedure; Carries out protCluster multiple
  times to assess consistency of assignments to compartments
}
