% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protDataPrep.r
\name{protDataPrep}
\alias{protDataPrep}
\title{Prepare data for constrained proportional assignment by reweighting differential and nyc fractions}
\usage{
protDataPrep(geneProfileSummary, n.diff.chan = 6, n.nyc.chan = 3,
  nycL1Adjust = T)
}
\arguments{
\item{geneProfileSummary}{data frame of gene/protein identifiers (geneName) and their relative abundance in centrifugation fractions.}

\item{n.diff.chan}{Number of channels of differential centrifugation fractions}

\item{n.nyc.chan}{Number of nycodenz fractions (derived from the L1 fraction)}

\item{nycL1Adjust}{True if adjust nyc by L1}
}
\value{
A matrix geneProfileSummary with re-weighted relative abundance fractions
}
\description{
Prepare data for constrained proportional assignment by reweighting differential and nyc fractions
}
