% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneProfilePlot.R
\name{protPlotfun}
\alias{protPlotfun}
\title{Plot profiles of reference genes}
\usage{
protPlotfun(
  protPlot,
  geneProfileSummary = geneProfileSummaryUse,
  finalList = finalListUse,
  n.fractions = 9,
  Nspectra = T,
  matLocR = matLocRuse,
  confInt = F,
  predictedPropL.mat = NULL,
  predictedPropU.mat
)
}
\arguments{
\item{geneProfileSummary}{data frame of protein names and relative abundance levels.}

\item{finalList}{spectrum-level abundance levels by protein and peptide}

\item{n.fractions}{number of fractions per protein}

\item{Nspectra}{indicator for if there are columns in geneProfileSummary for Nspectra
(number of spectra) and Nseq (number of peptides)}

\item{matLocR}{A matrix matLocR giving the abundance level profiles of the subcellular locations
8 columns are subcellular locations, and n.fractions rows are the fraction names}

\item{predictedPropL.mat}{Matrix of lower confidence limits, one row for each gene, and 9 columns
of lower confidence interval limits}

\item{predictedPropU.mat}{Matrix of upper confidence limits, one row for each gene, and 9 columns
of upper confidence interval limits}

\item{refLocProteins}{List of reference proteins}

\item{confint}{indicator for if there are standard errors (from bootstrapping) for the assigned proportions
(in geneProfileSummary)}
}
\description{
This function plots the average profiles of any gene in the dataset,
  the peptide profiles, and also the reference profile for each compartment
}
