% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protOutliersAndMeans.R,
%   R/protOutliersAndMeansOld.R
\name{readData}
\alias{readData}
\title{================================================================
readData to read in either cleaned data or outlier-identified
data, and assgin values for global variables (refCol, nCol, eps)}
\usage{
readData(dataUse, byPeptide, inputType = c("clean", "outlier"))

readData(dataUse, byPeptide, inputType = c("clean", "outlier"))
}
\arguments{
\item{dataUse:}{name of data set}

\item{byPeptide:}{identify outliers, and calculate mean
and SE by peptide or not
FALSE: by gene, TRUE: by pepetide}

\item{inputType:}{input cleaned data or outlier-identified
data}
}
\description{
================================================================
readData to read in either cleaned data or outlier-identified
data, and assgin values for global variables (refCol, nCol, eps)

================================================================
readData to read in either cleaned data or outlier-identified
data, and assgin values for global variables (refCol, nCol, eps)
}
