% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceProfilePlot.r
\name{referenceProfilePlot}
\alias{referenceProfilePlot}
\title{Plot profiles of reference genes}
\usage{
referenceProfilePlot(
  refLocProteins = refLocProteinsJadot,
  geneProfileSummary = geneProfileSummaryTMTms2,
  matLocR = matLocR,
  n.channels = 9
)
}
\arguments{
\item{refLocProteins}{List of reference proteins}

\item{geneProfileSummary}{data frame of protein names and relative abundance levels.}

\item{matLocR}{A matrix matLocR giving the abundance level profiles of the subcellular locations}
}
\description{
This function plots profiles of reference genes and also the average profile for each compartment
}
