% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refProfilePlot.r
\name{refProfilePlot}
\alias{refProfilePlot}
\title{Plot profiles of reference genes}
\usage{
refProfilePlot(
  refLoc,
  refLocProteins = refLocProteinsJadot,
  geneProfileSummary = geneProfileSummaryTMTms2,
  markerLocR = markerLocR,
  refGenePlot = NULL
)
}
\arguments{
\item{refLoc}{the name of the reference subcellular compartment to plot}

\item{refLocProteins}{List of reference proteins}

\item{geneProfileSummary}{data frame of protein names and relative abundance levels.}

\item{markerLocR}{A matrix markerLocR giving the abundance level profiles of the subcellular locations}
}
\description{
This function plots profiles of reference genes and also the average profile for each compartment
}
