% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAtransform.r
\name{relAmtTransform}
\alias{relAmtTransform}
\title{Set up RSA (relative specific amount) profiles for constrained proportional assignment}
\usage{
relAmtTransform(NSA, NstartMaterialFractions = 6, totProt = NULL)
}
\arguments{
\item{NSA}{A matrix giving the specific amount or normalized specific amount
profiles, either marker profiles from 'cpaSetup' (which requires
normalized specific amounts as input), or a list of many protein profiles}

\item{NstartMaterialFractions}{Number of differential fractions, typically 6, for N, M, L1, L2, P, and S}

\item{totProt}{Total protein counts in each of the differential and nycodenz fractions; this is necessary to compute RSA's}
}
\value{
amtProtFrac: amount of protein in a given fraction

relAmtProtFrac amount of given protein in fraction / amount of given protein in starting material
}
\description{
Set up RSA (relative specific amount) profiles for constrained proportional assignment
}
