% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAtransform.r
\name{RSAfromNSA}
\alias{RSAfromNSA}
\title{Directly compute relative specific activity from protProfileSummary (or from markerProfiles)}
\usage{
RSAfromNSA(
  NSA = protProfileLevels,
  NstartMaterialFractions = 6,
  totProt = NULL
)
}
\arguments{
\item{NstartMaterialFractions}{Number of differential fractions, typically 6, for N, M, L1, L2, P, and S}

\item{totProt}{Total protein counts in each of the fractions; this is necessary to compute RSA's}

\item{protProfileLevels}{A matrix (with no prot names) giving the abundance level profiles of the subcellular locations, from 'cpaSetup', or many protein profiles}
}
\value{
protProfileRSA: relative specific activity
}
\description{
Directly compute relative specific activity from protProfileSummary (or from markerProfiles)
}
