% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAabundanceTransform.R
\name{RSAtransform}
\alias{RSAtransform}
\title{Compute relative specific activity and RSA fractions from rlatve amounts in protein fractions}
\usage{
RSAtransform(
  relAmtProtFrac,
  nDiffFractions = 6,
  nNycFractions = 3,
  totProt = c(46.044776, 48.955954, 1.384083, 1.566324, 24.045584, 58.181818, 0.0368564,
    0.0684596, 1.27301587)
)
}
\arguments{
\item{relAmtProtFrac}{amount of given protein in fraction / amount of given protein in starting material}

\item{nDiffFractions}{Number of differential fractions, typically 6, for N, M, L1, L2, P, and S}

\item{nNycFractions}{Number of Nycodenz fractions, typically 3, but could be 1 (if only Nyc2 is present) or 0 if none}

\item{totProt}{Total protein counts in each of the differential and nycodenz fractions; this is necessary to compute RSA's}
}
\value{
rsa: relative specific activity
}
\description{
Compute relative specific activity and RSA fractions from rlatve amounts in protein fractions
}
