% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{cpaSetup}
\alias{cpaSetup}
\title{Set up reference profiles for constrained proportional assignment}
\usage{
cpaSetup(geneProfileSummary, refLocProteins = refLocProteins, n.channels)
}
\arguments{
\item{geneProfileSummary}{data frame of gene/protein identifiers (geneName) and their relative abundance in centrifugation fractions.}

\item{refLocProteins}{List of reference proteins and their subcellular locations}

\item{n.channels}{Number of channels of abundance levels}
}
\value{
A matrix matLocR giving the abundance level profiles of the subcellular locations
}
\description{
Set up reference profiles for constrained proportional assignment
}
