% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protOutliersAndMeans.R
\name{log2eps}
\alias{log2eps}
\title{================================================================
log2eps to log transformation with epsilon adjustment}
\usage{
log2eps(x, .eps = 0.029885209)
}
\arguments{
\item{x:}{a vector of numbers}

\item{eps:}{little value to avoid log2(0)}
}
\value{
: log transformed x
================================================================
}
\description{
================================================================
log2eps to log transformation with epsilon adjustment
}
