% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixturePlot.r
\name{mixturePlot}
\alias{mixturePlot}
\title{plot mixture of two compartment profiles}
\usage{
mixturePlot(
  mixProtiProtjCPA,
  NstartMaterialFractions = 6,
  Loc1,
  Loc2,
  input.prop,
  errorReturn = F,
  subTitle = NULL,
  xaxisLab = T,
  yaxisLab = T
)
}
\arguments{
\item{mixProtiProtjCPA}{data frame of CPA estimated proportions for each mixture}

\item{NstartMaterialFractions}{Number of fractions that comprise the starting material}

\item{Loc1}{row number of subcellular location 1 of mixture}

\item{Loc2}{row number of subcellular location 2 of mixture}

\item{input.prop}{true mixture proportions used to make the  mixtures (from proteinMix)}

\item{errorReturn}{Return area of error region if true}
}
\description{
plot mixture of two compartment profiles
}
