% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{proLocAll}
\alias{proLocAll}
\title{Carry out constrained proportional assignment on all proteinss}
\usage{
proLocAll(
  protProfileSummary,
  markerLocR,
  n.channels = n.channels,
  log2Transf = F,
  eps = 2^(-5),
  maxit = 10000,
  assignProbsStart = NULL,
  maxR = NULL
)
}
\arguments{
\item{protProfileSummary}{data frame of protein names and relative abundance levels.}

\item{markerLocR}{A matrix giving the abundance level profiles of the subcellular locations}

\item{n.channels}{Number of channels of abundance levels}

\item{log2Transf}{Transform data before running CPA? Default is F}

\item{eps}{Small quantity to add to data prior to taking a log transformation; ignored
if log2Transf=F}

\item{maxit}{maximum number of iterations}

\item{assignProbsStart}{A matrix of starting values, one for each protein. The first column must be protName}

\item{maxR}{A value to truncate values of markerLocR; if NULL (default), ignore}
}
\value{
assignProbsOut  Data frame of proportionate assignments of each protein to compartments
}
\description{
Carry out constrained proportional assignment on all proteinss
}
