% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{proLocAll}
\alias{proLocAll}
\title{Carry out constrained proportional assignment on all genes}
\usage{
proLocAll(geneProfileSummary, matLocR, n.channels = n.channels,
  log2Transf = F)
}
\arguments{
\item{geneProfileSummary}{data frame of protein names and relative abundance levels.}

\item{matLocR}{A matrix giving the abundance level profiles of the subcellular locations}

\item{n.channels}{Number of channels of abundance levels}
}
\value{
assignProbsOut  Data frame of proportionate assignments of each protein to compartments
}
\description{
Carry out constrained proportional assignment on all genes
}
