% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protProfilePlot.R
\name{protPlotfun}
\alias{protPlotfun}
\title{Plot profiles of reference proteins}
\usage{
protPlotfun(
  protName,
  profile,
  Nspectra = T,
  finalList = NULL,
  numDataCols = 9,
  n.compartments = 8,
  refLocationProfiles,
  assignPropsMat,
  propCI = F,
  transType = "",
  yAxisLabel = ""
)
}
\arguments{
\item{protName}{Name of the protein to plot}

\item{Nspectra}{indicator for if there are columns in profile for Nspectra
(number of spectra) and Npep (number of peptides)}

\item{finalList}{spectrum-level abundance levels by protein and peptide; Ehis is NULL
if not available}

\item{numDataCols}{number of fractions per protein}

\item{n.compartments}{number of compartments (8 in Jadot data)}

\item{refLocationProfiles}{A matrix refLocationProfiles giving the abundance level profiles of the subcellular locations
n.compartments = 8 columns are subcellular locations, and numDataCols rows are the fraction names}

\item{assignPropsMat}{A matrix of assignment proportions, from the constrained proportional assignment algorithm,
and optionally upper and lower 95 percent confidence limits}

\item{propCI}{True if lower and upper confidence intervals are included in assignPros}

\item{markerList}{List of reference proteins}

\item{protProfileSummary}{data frame of protein names and their relative abundance levels..}

\item{predictedProp.mat}{A matrix of CPA predicted proportions, from proLocAll}

\item{confint}{indicator for if there are standard errors (from bootstrapping) for the assigned proportions
(in profile)}

\item{predictedPropL.mat}{Matrix of lower confidence limits, one row for each protein, and 9 columns
of lower confidence interval limits}

\item{predictedPropU.mat}{Matrix of upper confidence limits, one row for each protein, and 9 columns
of upper confidence interval limits}
}
\description{
This function plots the average profiles of any protein in the dataset,
  the peptide profiles, and also the reference profile for each compartment
}
