% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneProfilePlot.R
\name{protPlotfun}
\alias{protPlotfun}
\title{Plot profiles of reference genes}
\usage{
protPlotfun(
  protName,
  geneProfileSummary = geneProfileSummaryUse,
  Nspectra = T,
  finalList = finalListUse,
  n.fractions = 9,
  n.compartments = 8,
  matLocR = matLocRuse,
  assignPropsMat = assignPropsUse,
  propCI = F
)
}
\arguments{
\item{protName}{Name of the protein (gene) to plot}

\item{geneProfileSummary}{data frame of protein names and their relative abundance levels..}

\item{Nspectra}{indicator for if there are columns in geneProfileSummary for Nspectra
(number of spectra) and Nseq (number of peptides)}

\item{finalList}{spectrum-level abundance levels by protein and peptide}

\item{n.fractions}{number of fractions per protein}

\item{n.compartments}{number of compartments (8 in Jadot data)}

\item{matLocR}{A matrix matLocR giving the abundance level profiles of the subcellular locations
n.compartments = 8 columns are subcellular locations, and n.fractions rows are the fraction names}

\item{assignPropsMat}{A matrix of assignment proportions, from the constrained proportional assignment algorithm,
and optionally upper and lower 95 percent confidence limits}

\item{propCI}{True if lower and upper confidence intervals are included in assignPros}

\item{refLocProteins}{List of reference proteins}

\item{predictedProp.mat}{A matrix of CPA predicted proportions, from proLocAll}

\item{confint}{indicator for if there are standard errors (from bootstrapping) for the assigned proportions
(in geneProfileSummary)}

\item{predictedPropL.mat}{Matrix of lower confidence limits, one row for each gene, and 9 columns
of lower confidence interval limits}

\item{predictedPropU.mat}{Matrix of upper confidence limits, one row for each gene, and 9 columns
of upper confidence interval limits}
}
\description{
This function plots the average profiles of any gene in the dataset,
  the peptide profiles, and also the reference profile for each compartment
}
