% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protProfilePlot.R
\name{protPlotfun}
\alias{protPlotfun}
\title{Plot profiles of reference proteins}
\usage{
protPlotfun(
  protName,
  protProfileSummary,
  Nspectra = T,
  finalList = NULL,
  n.fractions = 9,
  n.compartments = 8,
  markerLocR = markerLocRuse,
  assignPropsMat = assignPropsUse,
  propCI = F,
  transType = ""
)
}
\arguments{
\item{protName}{Name of the protein to plot}

\item{protProfileSummary}{data frame of protein names and their relative abundance levels..}

\item{Nspectra}{indicator for if there are columns in protProfileSummary for Nspectra
(number of spectra) and Nseq (number of peptides)}

\item{finalList}{spectrum-level abundance levels by protein and peptide; Ehis is NULL
if not available}

\item{n.fractions}{number of fractions per protein}

\item{n.compartments}{number of compartments (8 in Jadot data)}

\item{markerLocR}{A matrix markerLocR giving the abundance level profiles of the subcellular locations
n.compartments = 8 columns are subcellular locations, and n.fractions rows are the fraction names}

\item{assignPropsMat}{A matrix of assignment proportions, from the constrained proportional assignment algorithm,
and optionally upper and lower 95 percent confidence limits}

\item{propCI}{True if lower and upper confidence intervals are included in assignPros}

\item{refLocProteins}{List of reference proteins}

\item{predictedProp.mat}{A matrix of CPA predicted proportions, from proLocAll}

\item{confint}{indicator for if there are standard errors (from bootstrapping) for the assigned proportions
(in protProfileSummary)}

\item{predictedPropL.mat}{Matrix of lower confidence limits, one row for each protein, and 9 columns
of lower confidence interval limits}

\item{predictedPropU.mat}{Matrix of upper confidence limits, one row for each protein, and 9 columns
of upper confidence interval limits}
}
\description{
This function plots the average profiles of any protein in the dataset,
  the peptide profiles, and also the reference profile for each compartment
}
