% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAabundanceTransform.R
\name{rsaDirect}
\alias{rsaDirect}
\title{Directly compute relative specific activity from protProfileSummary (or from markerLocR)}
\usage{
rsaDirect(
  protProfileLevels,
  nDiffFractions = 6,
  nNycFractions = 3,
  totProt = c(46.044776, 48.955954, 1.384083, 1.566324, 24.045584, 58.181818,
    0.0368564, 0.0684596, 1.27301587),
  maxRSA = 22
)
}
\arguments{
\item{protProfileLevels}{A matrix (with no prot names) giving the abundance level profiles of the subcellular locations, from 'cpaSetup', or many protein profiles}

\item{nDiffFractions}{Number of differential fractions, typically 6, for N, M, L1, L2, P, and S}

\item{nNycFractions}{Number of Nycodenz fractions, typically 3, but could be 1 (if only Nyc2 is present) or 0 if none}

\item{totProt}{Total protein counts in each of the differential and nycodenz fractions; this is necessary to compute RSA's}
}
\value{
protProfileRSA: relative specific activity
}
\description{
Directly compute relative specific activity from protProfileSummary (or from markerLocR)
}
