% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protOutliersAndMeans.R,
%   R/protOutliersAndMeansOld.R
\name{scoresMod}
\alias{scoresMod}
\title{================================================================
scoresMod to identify outliers in a vector
Outliers as exceeding the normal probability of proba}
\usage{
scoresMod(x, reject.vec = NULL, proba = 0.99)

scoresMod(x, reject.vec = NULL, proba = 0.99)
}
\arguments{
\item{x:}{a vector of numbers}

\item{reject.vec.i:}{pre-specified rejection locations}

\item{proba:}{pre-specified normal probability}
}
\description{
================================================================
scoresMod to identify outliers in a vector
Outliers as exceeding the normal probability of proba

================================================================
scoresMod to identify outliers in a vector
Outliers as exceeding the normal probability of proba
}
\note{
: when sd(x)=0, scores() generates NA, mark
                      NA as FALSE
================================================================

: when sd(x)=0, scores() generates NA, mark
                      NA as FALSE
================================================================
}
